Add progressive hookshot

+++ include/combo/save.h
@@ -7,10 +7,10 @@
 /* Combo-specific save data */
 typedef struct PACKED ALIGNED(4)
 {
-    u32     ootTradeChild;
-    u32     ootTradeAdult;
-    u32     ootOcarinas;
-    u32     ootHookshots;
+    u16     ootTradeChild;
+    u16     ootTradeAdult;
+    u8      ootOcarinas;
+    u8      ootHookshots;
 }
 ComboSave;
